// packages/core/src/generated-core/models/Album.ts
/**
 * core
 * MSR Soundtrack audio management website\'s schema
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { FileStatus } from '../models/FileStatus.ts';
import { Song } from '../models/Song.ts';
import { HttpFile } from '../http/http.ts';

export class Album {
    'cid': string;
    'name': string;
    'songs': Array<Song>;
    'status': FileStatus;
    'coverPath': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "cid",
            "baseName": "cid",
            "type": "string",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "songs",
            "baseName": "songs",
            "type": "Array<Song>",
            "format": ""
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "FileStatus",
            "format": ""
        },
        {
            "name": "coverPath",
            "baseName": "coverPath",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Album.attributeTypeMap;
    }

    public constructor() {
    }
}