// packages/core/src/generated-core/models/FileStatus.ts
/**
 * Operation real wave API
 * MSR Soundtrack audio management website\'s schema
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http.ts';

export enum FileStatus {
    NotExists = 'NOT_EXISTS',
    Broken = 'BROKEN',
    Exists = 'EXISTS'
}